Refactor time pivot and update display
