+++ src/Client/features/plotlyChartProvider.ts
@@ -6,9 +6,9 @@
  * Ported from Server/Charting C# implementation.
  */
 
-import type { ChartOptions, ResultColumn, ResultTable } from './server';
-import { ChartType, ChartKind, ChartAxis, ChartSortOrder, ChartLegendPosition } from './chartProvider';
-import type { IChartView, IWebView, IChartProvider } from './chartProvider';
+import type { ChartOptions, ResultTable } from './server';
+import { ChartType, ChartKind, ChartAxis, ChartSortOrder, ChartLegendPosition, ChartColorways, hexToRgba, isNumericType, isDateTimeType, getColumnRef, getColumnRefByIndex } from './chartProvider';
+import type { IChartView, IWebView, IChartProvider, ColumnRef } from './chartProvider';
 
 // ─── Plotly Constants ───────────────────────────────────────────────────────
 
@@ -59,12 +59,7 @@ const PlotlyAxisSides = {
     Right: 'right',
 } as const;
 
-const PlotlyColorways = {
-    Default: [
-        '#636EFA', '#EF553B', '#00CC96', '#AB63FA', '#FDC826',
-        '#19D3F3', '#FF6692', '#B6E880', '#FF97FF', '#FECB52',
-    ],
-} as const;
+const PlotlyColorways = ChartColorways;
 
 const SankeyDefaultNodeColors: readonly string[] = [
     '#1f77b4', '#ff7f0e', '#2ca02c', '#d62728', '#9467bd',
@@ -918,36 +913,6 @@ try {
 
 // ─── Utility Functions ──────────────────────────────────────────────────────
 
-function hexToRgba(hex: string, opacity: number): string {
-    let h = hex.replace(/^#/, '');
-    if (h.length === 3) {
-        h = (h[0] ?? '') + (h[0] ?? '') + (h[1] ?? '') + (h[1] ?? '') + (h[2] ?? '') + (h[2] ?? '');
-    }
-    if (h.length !== 6) {
-        return `rgba(128, 128, 128, ${opacity})`;
-    }
-    const r = parseInt(h.substring(0, 2), 16);
-    const g = parseInt(h.substring(2, 4), 16);
-    const b = parseInt(h.substring(4, 6), 16);
-    return `rgba(${r}, ${g}, ${b}, ${opacity})`;
-}
-
-function isNumericType(type: string): boolean {
-    switch (type) {
-        case 'int':
-        case 'long':
-        case 'real':
-        case 'decimal':
-            return true;
-        default:
-            return false;
-    }
-}
-
-function isDateTimeType(type: string): boolean {
-    return type === 'datetime' || type === 'timespan';
-}
-
 function isTimeChartType(type: string): boolean {
     return type === ChartType.TimeLineChart
         || type === ChartType.TimeLineWithAnomalyChart
@@ -1128,28 +1093,7 @@ function tryGetDouble(value: unknown): number | undefined {
     return n;
 }
 
-// ─── Data Access Helpers (replaces DataTable/DataColumn) ────────────────────
-
-/** A ResultColumn together with its position in the table. */
-interface ColumnRef {
-    column: ResultColumn;
-    index: number;
-}
-
-function getColumnRef(table: ResultTable, name: string): ColumnRef | undefined {
-    const idx = table.columns.findIndex(c => c.name === name);
-    if (idx < 0) return undefined;
-    const col = table.columns[idx];
-    if (!col) return undefined;
-    return { column: col, index: idx };
-}
-
-function getColumnRefByIndex(table: ResultTable, index: number): ColumnRef | undefined {
-    if (index < 0 || index >= table.columns.length) return undefined;
-    const col = table.columns[index];
-    if (!col) return undefined;
-    return { column: col, index };
-}
+// ─── Data Access Helpers ────────────────────────────────────────────────────
 
 function getColumnValues(table: ResultTable, col: ColumnRef): unknown[] {
     return table.rows.map(row => row?.[col.index]);
@@ -1461,10 +1405,6 @@ export class PlotlyChartProvider implements IChartProvider {
             return this.renderRawPlotlyChart(data, darkMode);
         }
 
-        if (options.type === ChartType.TimePivot) {
-            return this.renderTimePivotHtml(data, options, darkMode);
-        }
-
         const builder = this.buildChart(data, options, darkMode);
         return builder?.toHtmlDiv();
     }
@@ -1546,10 +1486,8 @@ export class PlotlyChartProvider implements IChartProvider {
                 builder = this.buildSankeyChart(data, options);
                 break;
             case ChartType.TimeLadderChart:
-                builder = this.buildLadderChart(data, options);
-                break;
             case ChartType.TimePivot:
-                builder = this.buildTimePivotChart(data, options);
+                builder = this.buildLadderChart(data, options);
                 break;
             default:
                 return undefined;
@@ -2032,683 +1970,6 @@ export class PlotlyChartProvider implements IChartProvider {
         builder = applyCommonOptions(builder, options);
 
         return builder;
-    }
-
-    // ─── Time Pivot Chart ───────────────────────────────────────────────
-
-    private buildTimePivotChart(data: ResultTable, options: ChartOptions): PlotlyChartBuilder | undefined {
-        if (data.columns.length < 2 || data.rows.length === 0) return undefined;
-
-        // Find datetime columns and candidate series columns
-        const datetimeCols: ColumnRef[] = [];
-        const seriesCandidates: ColumnRef[] = [];
-
-        for (let i = 0; i < data.columns.length; i++) {
-            const col = data.columns[i];
-            if (!col) continue;
-            if (isDateTimeType(col.type)) {
-                const ref = getColumnRefByIndex(data, i);
-                if (ref) datetimeCols.push(ref);
-            } else if (!isNumericType(col.type)) {
-                const ref = getColumnRefByIndex(data, i);
-                if (ref) seriesCandidates.push(ref);
-            }
-        }
-
-        if (datetimeCols.length === 0) return undefined;
-
-        const startColumn = datetimeCols[0]!;
-        const endColumn = datetimeCols.length >= 2 ? datetimeCols[1]! : undefined;
-        const isRangeMode = endColumn !== undefined;
-
-        // Resolve series columns: use explicit, or auto-infer first non-datetime non-numeric
-        let seriesCols: ColumnRef[];
-        if (options.seriesColumns && options.seriesColumns.length > 0) {
-            seriesCols = options.seriesColumns
-                .map(name => getColumnRef(data, name))
-                .filter((c): c is ColumnRef => c !== undefined);
-        } else {
-            seriesCols = seriesCandidates.length > 0 ? [seriesCandidates[0]!] : [];
-        }
-
-        if (seriesCols.length === 0) return undefined;
-
-        const primaryCol = seriesCols[0]!;
-        const secondaryCols = seriesCols.slice(1);
-        const hasNesting = secondaryCols.length > 0;
-
-        // Group data: primaryKey → secondaryKey → rows
-        interface PivotRow { start: number; end: number; }
-        const groups = new Map<string, Map<string, PivotRow[]>>();
-        const primaryOrder: string[] = [];
-
-        for (const row of data.rows) {
-            if (!row) continue;
-            const startVal = row[startColumn.index];
-            if (startVal == null) continue;
-            const startMs = new Date(String(startVal)).getTime();
-            if (isNaN(startMs)) continue;
-
-            let endMs: number;
-            if (isRangeMode) {
-                const endVal = row[endColumn.index];
-                if (endVal == null) continue;
-                endMs = new Date(String(endVal)).getTime();
-                if (isNaN(endMs)) continue;
-            } else {
-                endMs = startMs; // point mode: start == end
-            }
-
-            const primaryKey = String(row[primaryCol.index] ?? '');
-            const secondaryKey = hasNesting
-                ? secondaryCols.map(c => String(row[c.index] ?? '')).join(' - ')
-                : primaryKey;
-
-            if (!groups.has(primaryKey)) {
-                groups.set(primaryKey, new Map());
-                primaryOrder.push(primaryKey);
-            }
-            const secondaryMap = groups.get(primaryKey)!;
-            if (!secondaryMap.has(secondaryKey)) {
-                secondaryMap.set(secondaryKey, []);
-            }
-            secondaryMap.get(secondaryKey)!.push({ start: startMs, end: endMs });
-        }
-
-        if (primaryOrder.length === 0) return undefined;
-
-        // Build y-axis categories in display order (top-to-bottom)
-        const categories: string[] = [];
-        for (const primaryKey of primaryOrder) {
-            if (hasNesting) categories.push(primaryKey);
-            const secondaryMap = groups.get(primaryKey)!;
-            for (const secondaryKey of secondaryMap.keys()) {
-                categories.push(hasNesting ? `   ${secondaryKey}` : secondaryKey);
-            }
-        }
-
-        let builder = new PlotlyChartBuilder();
-        if (options.title) builder = builder.withTitle(options.title);
-        if (options.xTitle) builder = builder.setXAxisTitle(options.xTitle);
-        if (options.yTitle) builder = builder.setYAxisTitle(options.yTitle);
-        if (options.showLegend === false) builder = builder.hideLegend();
-        builder = applyCommonOptions(builder, options);
-
-        builder = builder.setBarMode('overlay');
-        builder = builder.withXAxis({ ...(builder.layout.xaxis ?? {}), type: 'date' });
-        builder = builder.withYAxis({
-            ...(builder.layout.yaxis ?? {}),
-            categoryorder: 'array' as const,
-            categoryarray: [...categories].reverse(),
-        });
-        builder = builder.withLayout({
-            ...builder.layout,
-            legend: { ...(builder.layout.legend ?? {}), traceorder: 'reversed' as const },
-        });
-
-        const colors = PlotlyColorways.Default;
-        const segmentLineWidth = 8;
-        const bgBarOpacity = 0.2;
-        const parentBarOpacity = 0.3;
-
-        let colorIndex = 0;
-        for (const primaryKey of primaryOrder) {
-            const color = colors[colorIndex % colors.length]!;
-            const bgColor = hexToRgba(color, bgBarOpacity);
-            const parentColor = hexToRgba(color, parentBarOpacity);
-            const secondaryMap = groups.get(primaryKey)!;
-            const legendGroup = primaryKey;
-
-            // Collect ranges across all secondaries for the parent bar
-            let groupMinStart = Infinity;
-            let groupMaxEnd = -Infinity;
-            for (const rows of secondaryMap.values()) {
-                for (const r of rows) {
-                    if (r.start < groupMinStart) groupMinStart = r.start;
-                    if (r.end > groupMaxEnd) groupMaxEnd = r.end;
-                }
-            }
-
-            // Parent background bar (only with nesting)
-            if (hasNesting) {
-                const duration = groupMaxEnd - groupMinStart;
-                builder = builder.addTrace({
-                    type: 'bar',
-                    x: [duration],
-                    y: [primaryKey],
-                    base: [new Date(groupMinStart).toISOString()],
-                    orientation: PlotlyOrientations.Horizontal,
-                    marker: { color: parentColor },
-                    name: primaryKey,
-                    legendgroup: legendGroup,
-                    showlegend: true,
-                    hoverinfo: 'name',
-                } as BarTrace);
-            }
-
-            // Detail rows
-            for (const [secondaryKey, rows] of secondaryMap) {
-                const yLabel = hasNesting ? `   ${secondaryKey}` : secondaryKey;
-
-                // Detail background bar: spans min-to-max of this secondary's data
-                let minStart = Infinity;
-                let maxEnd = -Infinity;
-                for (const r of rows) {
-                    if (r.start < minStart) minStart = r.start;
-                    if (r.end > maxEnd) maxEnd = r.end;
-                }
-
-                if (isRangeMode) {
-                    const bgDuration = maxEnd - minStart;
-                    builder = builder.addTrace({
-                        type: 'bar',
-                        x: [bgDuration],
-                        y: [yLabel],
-                        base: [new Date(minStart).toISOString()],
-                        orientation: PlotlyOrientations.Horizontal,
-                        marker: { color: bgColor },
-                        name: hasNesting ? secondaryKey : primaryKey,
-                        legendgroup: legendGroup,
-                        showlegend: !hasNesting,
-                        hoverinfo: 'name',
-                    } as BarTrace);
-
-                    // Segment lines: thick lines for each range
-                    const segX: unknown[] = [];
-                    const segY: unknown[] = [];
-                    for (const r of rows) {
-                        segX.push(new Date(r.start).toISOString(), new Date(r.end).toISOString(), null);
-                        segY.push(yLabel, yLabel, null);
-                    }
-                    builder = builder.addTrace({
-                        type: 'scatter',
-                        x: segX,
-                        y: segY,
-                        mode: PlotlyScatterModes.Lines,
-                        line: { color, width: segmentLineWidth },
-                        name: hasNesting ? secondaryKey : primaryKey,
-                        legendgroup: legendGroup,
-                        showlegend: false,
-                        hoverinfo: 'x+name',
-                    } as ScatterTrace);
-                } else {
-                    // Point mode: scatter markers at each datetime
-                    const ptX = rows.map(r => new Date(r.start).toISOString());
-                    const ptY = rows.map(() => yLabel);
-                    builder = builder.addTrace({
-                        type: 'scatter',
-                        x: ptX,
-                        y: ptY,
-                        mode: PlotlyScatterModes.Markers,
-                        marker: { color, size: 8 },
-                        name: hasNesting ? secondaryKey : primaryKey,
-                        legendgroup: legendGroup,
-                        showlegend: !hasNesting,
-                        hoverinfo: 'x+name',
-                    } as ScatterTrace);
-                }
-            }
-
-            colorIndex++;
-        }
-
-        return builder;
-    }
-
-    // ─── Time Pivot HTML Renderer ───────────────────────────────────────
-
-    private renderTimePivotHtml(data: ResultTable, options: ChartOptions, darkMode: boolean): string | undefined {
-        if (data.columns.length < 2 || data.rows.length === 0) return undefined;
-
-        // Find datetime columns and candidate series columns
-        const datetimeCols: ColumnRef[] = [];
-        const seriesCandidates: ColumnRef[] = [];
-
-        for (let i = 0; i < data.columns.length; i++) {
-            const col = data.columns[i];
-            if (!col) continue;
-            if (isDateTimeType(col.type)) {
-                const ref = getColumnRefByIndex(data, i);
-                if (ref) datetimeCols.push(ref);
-            } else if (!isNumericType(col.type)) {
-                const ref = getColumnRefByIndex(data, i);
-                if (ref) seriesCandidates.push(ref);
-            }
-        }
-
-        if (datetimeCols.length === 0) return undefined;
-
-        const startColumn = datetimeCols[0]!;
-        const endColumn = datetimeCols.length >= 2 ? datetimeCols[1]! : undefined;
-        const isRangeMode = endColumn !== undefined;
-
-        // Resolve series columns: use explicit, or auto-infer first non-datetime non-numeric
-        let seriesCols: ColumnRef[];
-        if (options.seriesColumns && options.seriesColumns.length > 0) {
-            seriesCols = options.seriesColumns
-                .map(name => getColumnRef(data, name))
-                .filter((c): c is ColumnRef => c !== undefined);
-        } else {
-            seriesCols = seriesCandidates.length > 0 ? [seriesCandidates[0]!] : [];
-        }
-
-        if (seriesCols.length === 0) return undefined;
-
-        // Build tree: each series column = one nesting level, leaf = data rows
-        interface PivotRow { start: number; end: number; }
-        interface TreeNode {
-            key: string;
-            children: Map<string, TreeNode>;
-            childOrder: string[];
-            rows: PivotRow[];
-            level: number;
-        }
-
-        const root: TreeNode = { key: '', children: new Map(), childOrder: [], rows: [], level: -1 };
-        let globalMin = Infinity;
-        let globalMax = -Infinity;
-
-        for (const row of data.rows) {
-            if (!row) continue;
-            const startVal = row[startColumn.index];
-            if (startVal == null) continue;
-            const startMs = new Date(String(startVal)).getTime();
-            if (isNaN(startMs)) continue;
-
-            let endMs: number;
-            if (isRangeMode) {
-                const endVal = row[endColumn.index];
-                if (endVal == null) continue;
-                endMs = new Date(String(endVal)).getTime();
-                if (isNaN(endMs)) continue;
-            } else {
-                endMs = startMs;
-            }
-
-            if (startMs < globalMin) globalMin = startMs;
-            if (endMs > globalMax) globalMax = endMs;
-
-            // Walk down the tree, creating nodes at each series level
-            let node = root;
-            for (let lvl = 0; lvl < seriesCols.length; lvl++) {
-                const key = String(row[seriesCols[lvl]!.index] ?? '');
-                if (!node.children.has(key)) {
-                    node.children.set(key, { key, children: new Map(), childOrder: [], rows: [], level: lvl });
-                    node.childOrder.push(key);
-                }
-                node = node.children.get(key)!;
-            }
-            node.rows.push({ start: startMs, end: endMs });
-        }
-
-        if (root.childOrder.length === 0 || globalMin >= globalMax) return undefined;
-
-        const globalRange = globalMax - globalMin;
-        const colors = PlotlyColorways.Default;
-        const maxLevel = seriesCols.length - 1;
-
-        const pct = (ms: number) => ((ms - globalMin) / globalRange * 100);
-        const esc = (s: string) => s.replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;').replace(/"/g, '&quot;');
-
-        // Compute time range for a node (all descendant rows)
-        function nodeRange(node: TreeNode): { min: number; max: number } {
-            let min = Infinity;
-            let max = -Infinity;
-            for (const r of node.rows) {
-                if (r.start < min) min = r.start;
-                if (r.start > max) max = r.start;
-                if (r.end < min) min = r.end;
-                if (r.end > max) max = r.end;
-            }
-            for (const child of node.children.values()) {
-                const cr = nodeRange(child);
-                if (cr.min < min) min = cr.min;
-                if (cr.max > max) max = cr.max;
-            }
-            return { min, max };
-        }
-
-        // Count leaf descendants
-        function leafCount(node: TreeNode): number {
-            if (node.children.size === 0) return 1;
-            let count = 0;
-            for (const child of node.children.values()) count += leafCount(child);
-            return count;
-        }
-
-        // Collect all rows from descendant leaves
-        function allRows(node: TreeNode): PivotRow[] {
-            if (node.children.size === 0) return node.rows;
-            const result: PivotRow[] = [];
-            for (const child of node.children.values()) result.push(...allRows(child));
-            return result;
-        }
-
-        // Build time axis ticks
-        const tickCount = 8;
-        const ticksHtml: string[] = [];
-        for (let i = 0; i <= tickCount; i++) {
-            const ms = globalMin + (globalRange * i / tickCount);
-            const d = new Date(ms);
-            const label = d.toISOString().replace('T', ' ').replace(/\.000Z$/, '');
-            const left = (i / tickCount * 100);
-            ticksHtml.push(`<div class="tp-tick" style="left:${left}%"><div class="tp-tick-line"></div><div class="tp-tick-label">${esc(label)}</div></div>`);
-        }
-
-        // Recursively build rows
-        const rowsHtml: string[] = [];
-        let colorIndex = 0;
-
-        function renderNode(node: TreeNode, color: string, depth: number, parentPath: string): void {
-            const isLeaf = node.level === maxLevel;
-            const indent = depth * 16;
-            const range = nodeRange(node);
-            const bgLeft = pct(range.min);
-            const bgWidth = pct(range.max) - bgLeft;
-            const nodeRange_ = range.max - range.min;
-            const pctLocal = (ms: number) => nodeRange_ > 0 ? ((ms - range.min) / nodeRange_ * 100) : 0;
-
-            if (!isLeaf) {
-                // Group row with toggle + spanning bar
-                const count = leafCount(node);
-                const bgOpacity = 0.3 - depth * 0.05; // slightly lighter at each level
-                const nodePath = parentPath ? `${parentPath}/${node.key}` : node.key;
-                const collapsedClass = 'collapsed';
-                const hiddenClass = depth > 0 ? ' tp-hidden' : '';
-
-                // Build aggregated segments from all descendant leaves
-                let aggSegments = '';
-                const descRows = allRows(node);
-                if (isRangeMode) {
-                    for (const r of descRows) {
-                        const segLeft = pctLocal(r.start);
-                        const segWidth = pctLocal(r.end) - segLeft;
-                        const startLabel = new Date(r.start).toISOString().replace('T', ' ').replace(/\.000Z$/, '');
-                        const endLabel = new Date(r.end).toISOString().replace('T', ' ').replace(/\.000Z$/, '');
-                        aggSegments += `<div class="tp-segment" style="left:${segLeft}%;width:${segWidth}%;background:${color}" title="${esc(startLabel)} → ${esc(endLabel)}"></div>`;
-                    }
-                } else {
-                    for (const r of descRows) {
-                        const dotLeft = pctLocal(r.start);
-                        const dotLabel = new Date(r.start).toISOString().replace('T', ' ').replace(/\.000Z$/, '');
-                        aggSegments += `<div class="tp-dot" style="left:${dotLeft}%;background:${color}" title="${esc(dotLabel)}"></div>`;
-                    }
-                }
-
-                const rangeStartLabel = new Date(range.min).toISOString().replace('T', ' ').replace(/\.000Z$/, '');
-                const rangeEndLabel = new Date(range.max).toISOString().replace('T', ' ').replace(/\.000Z$/, '');
-                const bgTitle = `${esc(node.key)}: ${esc(rangeStartLabel)} → ${esc(rangeEndLabel)} (${count})`;
-
-                rowsHtml.push(
-                    `<div class="tp-row tp-group ${collapsedClass}${hiddenClass}" data-group="${esc(nodePath)}" data-depth="${depth}">` +
-                    `<div class="tp-label tp-group-label" style="padding-left:${8 + indent}px">` +
-                    `<span class="tp-toggle">&#9662;</span>` +
-                    `${esc(node.key)} <span class="tp-count">(${count})</span></div>` +
-                    `<div class="tp-swimlane"><div class="tp-bg-bar" style="left:${bgLeft}%;width:${bgWidth}%;background:${hexToRgba(color, Math.max(bgOpacity, 0.1))}" title="${bgTitle}"><div class="tp-agg">${aggSegments}</div></div></div></div>`
-                );
-
-                // Render children
-                for (const childKey of node.childOrder) {
-                    const child = node.children.get(childKey)!;
-                    renderNode(child, color, depth + 1, nodePath);
-                }
-            } else {
-                // Leaf detail row
-                const bgColor = hexToRgba(color, 0.15);
-                let segmentsHtml = '';
-
-                if (isRangeMode) {
-                    for (const r of node.rows) {
-                        const segLeft = pctLocal(r.start);
-                        const segWidth = pctLocal(r.end) - segLeft;
-                        const startLabel = new Date(r.start).toISOString().replace('T', ' ').replace(/\.000Z$/, '');
-                        const endLabel = new Date(r.end).toISOString().replace('T', ' ').replace(/\.000Z$/, '');
-                        segmentsHtml += `<div class="tp-segment" style="left:${segLeft}%;width:${segWidth}%;background:${color}" title="${esc(startLabel)} → ${esc(endLabel)}"></div>`;
-                    }
-                } else {
-                    for (const r of node.rows) {
-                        const dotLeft = pctLocal(r.start);
-                        const dotLabel = new Date(r.start).toISOString().replace('T', ' ').replace(/\.000Z$/, '');
-                        segmentsHtml += `<div class="tp-dot" style="left:${dotLeft}%;background:${color}" title="${esc(dotLabel)}"></div>`;
-                    }
-                }
-
-                const parentPath2 = parentPath ? parentPath : '';
-                const leafHidden = depth > 0 ? ' tp-hidden' : '';
-                rowsHtml.push(
-                    `<div class="tp-row tp-leaf${leafHidden}" data-parent-group="${esc(parentPath2)}" data-depth="${depth}">` +
-                    `<div class="tp-label" style="padding-left:${8 + indent}px">${esc(node.key)}</div>` +
-                    `<div class="tp-swimlane">` +
-                    `<div class="tp-bg-bar" style="left:${bgLeft}%;width:${bgWidth}%;background:${bgColor}">` +
-                    segmentsHtml +
-                    `</div></div></div>`
-                );
-            }
-        }
-
-        // Render top-level nodes
-        for (const topKey of root.childOrder) {
-            const topNode = root.children.get(topKey)!;
-            const isTopLeaf = topNode.level === maxLevel;
-            const color = colors[colorIndex % colors.length]!;
-
-            if (isTopLeaf && seriesCols.length === 1) {
-                // Single series column: render flat (no group wrapper)
-                renderNode(topNode, color, 0, '');
-            } else {
-                renderNode(topNode, color, 0, '');
-            }
-            colorIndex++;
-        }
-
-        const titleHtml = options.title ? `<div class="tp-title">${esc(options.title)}</div>` : '';
-
-        return `
-<style>
-.tp-container {
-    font-family: var(--vscode-font-family, sans-serif);
-    font-size: var(--vscode-font-size, 13px);
-    color: var(--vscode-foreground);
-    overflow: auto;
-    height: 100%;
-    user-select: none;
-}
-.tp-title {
-    font-size: 1.2em;
-    font-weight: 600;
-    padding: 8px 12px 4px;
-}
-.tp-axis {
-    display: flex;
-    position: sticky;
-    top: 0;
-    z-index: 2;
-    background: var(--vscode-editor-background);
-    border-bottom: 1px solid var(--vscode-editorWidget-border, #444);
-}
-.tp-axis-label-spacer {
-    flex: 0 0 180px;
-    min-width: 180px;
-    position: sticky;
-    left: 0;
-    z-index: 3;
-    background: var(--vscode-editor-background);
-}
-.tp-axis-ticks {
-    flex: 1;
-    position: relative;
-    height: 36px;
-}
-.tp-tick {
-    position: absolute;
-    top: 0;
-    height: 100%;
-}
-.tp-tick-line {
-    width: 1px;
-    height: 18px;
-    background: var(--vscode-editorWidget-border, #555);
-}
-.tp-tick-label {
-    font-size: 10px;
-    white-space: nowrap;
-    color: var(--vscode-descriptionForeground, #999);
-    transform: translateX(-50%);
-    padding-top: 2px;
-}
-.tp-rows {
-    position: relative;
-}
-.tp-row {
-    display: flex;
-    align-items: center;
-    min-height: 26px;
-    border-bottom: 1px solid var(--vscode-editorWidget-border, rgba(128,128,128,0.2));
-}
-.tp-group {
-    min-height: 28px;
-}
-.tp-label {
-    flex: 0 0 180px;
-    min-width: 180px;
-    padding: 2px 8px;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    font-size: 12px;
-    position: sticky;
-    left: 0;
-    z-index: 1;
-    background: var(--vscode-editor-background);
-}
-.tp-group-label {
-    font-weight: 600;
-}
-.tp-count {
-    color: var(--vscode-descriptionForeground, #888);
-    font-weight: normal;
-    font-size: 11px;
-}
-.tp-toggle {
-    cursor: pointer;
-    display: inline-block;
-    transition: transform 0.15s;
-    margin-right: 4px;
-}
-.tp-group.collapsed .tp-toggle {
-    transform: rotate(-90deg);
-}
-.tp-agg {
-    position: absolute;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    pointer-events: none;
-}
-.tp-agg > * {
-    pointer-events: auto;
-}
-.tp-group:not(.collapsed) .tp-agg {
-    display: none;
-}
-.tp-hidden {
-    display: none !important;
-}
-.tp-swimlane {
-    flex: 1;
-    position: relative;
-    height: 20px;
-    margin: 2px 8px 2px 0;
-    overflow: hidden;
-}
-.tp-group .tp-swimlane {
-    height: 16px;
-}
-.tp-bg-bar {
-    position: absolute;
-    top: 0;
-    height: 100%;
-    border-radius: 3px;
-    min-width: 8px;
-    overflow: hidden;
-}
-.tp-segment {
-    position: absolute;
-    top: 50%;
-    height: 6px;
-    transform: translateY(-50%);
-    border-radius: 2px;
-    opacity: 0.9;
-    min-width: 8px;
-}
-.tp-dot {
-    position: absolute;
-    top: 50%;
-    width: 8px;
-    height: 8px;
-    border-radius: 50%;
-    transform: translate(-50%, -50%);
-}
-.tp-grid-line {
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    width: 1px;
-    background: var(--vscode-editorWidget-border, rgba(128,128,128,0.15));
-    pointer-events: none;
-}
-</style>
-<div class="tp-container">
-    ${titleHtml}
-    <div class="tp-axis">
-        <div class="tp-axis-label-spacer"></div>
-        <div class="tp-axis-ticks">${ticksHtml.join('')}</div>
-    </div>
-    <div class="tp-rows" id="tp-rows-container">
-        ${rowsHtml.join('\n        ')}
-    </div>
-</div>
-<script>
-(function() {
-    // Toggle collapse/expand for any group level
-    document.querySelectorAll('.tp-toggle').forEach(function(toggle) {
-        toggle.addEventListener('click', function(e) {
-            e.stopPropagation();
-            var group = this.closest('.tp-group');
-            if (!group) return;
-            var isCollapsing = !group.classList.contains('collapsed');
-            group.classList.toggle('collapsed');
-            var groupPath = group.dataset.group;
-            // Hide/show all descendant rows whose parent-group starts with this path
-            var allRows = group.parentElement.children;
-            var found = false;
-            for (var i = 0; i < allRows.length; i++) {
-                if (allRows[i] === group) { found = true; continue; }
-                if (!found) continue;
-                var row = allRows[i];
-                var rowGroup = row.dataset.group || row.dataset.parentGroup || '';
-                // Stop if we hit a sibling at same or higher level
-                if (row.classList.contains('tp-group') && !rowGroup.startsWith(groupPath + '/')) break;
-                if (row.classList.contains('tp-leaf') && !rowGroup.startsWith(groupPath)) break;
-                if (isCollapsing) {
-                    row.classList.add('tp-hidden');
-                    // Also collapse nested groups
-                    if (row.classList.contains('tp-group')) row.classList.add('collapsed');
-                } else {
-                    // Only show direct children (not nested collapsed ones)
-                    var rowDepth = parseInt(row.dataset.depth || '0');
-                    var groupDepth = parseInt(group.dataset.depth || '0');
-                    if (rowDepth === groupDepth + 1) {
-                        row.classList.remove('tp-hidden');
-                        // Keep nested groups collapsed
-                    }
-                }
-            }
-        });
-    });
-})();
-</script>`;
     }
 
     // ─── Ladder Chart ─────────────────────────────────────────────────
